main - use i16

+++ src/lib.rs
@@ -225,7 +225,7 @@ pub struct Stream {
 #[allow(clippy::derive_partial_eq_without_eq)]
 pub struct SideData {
     pub side_data_type: String,
-    pub rotation: Option<i8>,
+    pub rotation: Option<i16>,
 }
 
 #[derive(Default, Debug, Clone, PartialEq, serde::Serialize, serde::Deserialize)]
